patchwork: switch back to series instead of events
Events api in patchwork is not that straigthforward, the timestamps on
events are ~12 hours bigger than on corresponding series/patches for
some reason and events api calls on patchwork.ozlabs.org take forever
(sometimes timeouting). Commit 490ca16 "patchwork2: skip incomplete series"
seems to be enough to solve the problem of incomplete series, so
switching back to querying series instead of relying on events.