Allow extension of skip patterns
The default of this option is not useful for any other project than
netdev. Add option to allow extending it with additional patterns and
move initialization to parent class. Also clean up related code (move
constant list to the beginning of the file, rename self.skp to self.skip
to make the name more self-descriptive).

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>