Clear up skip logic for Patchwork2
For Patchwork2, we only check for matches in series name and don't check
patches themselves. This is a problem with series which for example
contain patches for both kernel and kernel-related tools that depend on
the kernel change.

Add skip pattern checking for patch names and clear up skip logic in
relevant docstrings.

Fixes #54

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>