Remove patchtest table
Data are inserted into the table, only to never be read. Results are
supposed to be reported to Patchwork and emails so there's no need to
keep the table. We still save tested patches to check for last tested
patch, but we don't need associated data.

Removal of the table allows us to simplify commit_patchtest method
(renamed to commit_tested) and delete it's test since no data is
committed in the method.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>