Handle Jenkins errors
Consider any result coming from Jenkins, except straight SUCCESS or
UNSTABLE (with matching skt results), an error. Ignore builds with such
results, leaving patch tests in "pendingpatches", and baseline untested,
both to be retried later.

UNSTABLE is a Jenkins build result signifying some errors, which were
not fatal for the completion of the build, while FAILURE is a fatal
build error. We can consider "build" to be our test run, and test
failures not fatal for the run. It's awkward naming, but it seems to be
the best mapping to Jenkins build results.

See also http://javadoc.jenkins-ci.org/hudson/model/Result.html