Composition fixups for _get_project_id
Introduce a method in the parent class that the child classes are
supposed to implement and override (fixes composition and pylint error)
and modify the argument name of the Patchwork1Project class' method to
match the one in the parent. Also fix the docstring about returned value
which didn't match what was actually being returned.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>