Unify get_patch_by_id
Throw an exception if the requested patch can't be retrieved. This was
already done for PatchworkV2Project but not PatchworkV1Project. Being
unable to retrieve a patch previously known as existing signals a
possible problem in sktm's logic, and it should be handled as such. The
unification also allows to remove the return value checks when this
method is used.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>